backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-decouple==3.8
Pillow==11.0.0
django-filter==23.3
python-dotenv
dj-database-url==1.2.0
